numeric dominance pruning

+++ src/search/numeric_dominance/breadth_first_search.h
@@ -4,11 +4,11 @@
 #include <vector>
 #include <limits>
 
-template <typename T> 
+template<typename T>
 void breadth_first_search_reachability_distances_one(const std::vector<std::vector<int> > &graph,
-						     int initial_state,
-						     std::vector<T> & distances,
-						     std::vector<int> & reachable_states) {
+                                                     int initial_state,
+                                                     std::vector<T> &distances,
+                                                     std::vector<int> &reachable_states) {
     int num_states = graph.size();
     distances.resize(num_states);
     std::fill(distances.begin(), distances.end(), std::numeric_limits<int>::max());
@@ -18,20 +18,19 @@ void breadth_first_search_reachability_distances_one(const std::vector<std::vect
 
     distances[initial_state] = 0;
     reachable_states.push_back(initial_state);
-    for(size_t i = 0; i < reachable_states.size(); ++i) {
-	if (increase_distance_in == i) {
-	    current_distance ++;
-	    increase_distance_in = reachable_states.size();
-	}
-	for (int successor : graph[reachable_states[i]]) {
-	    if (distances[successor] > current_distance) {
-		distances[successor] = current_distance;
-		reachable_states.push_back(successor);
-	    }
-	}
+    for (size_t i = 0; i < reachable_states.size(); ++i) {
+        if (size_t(increase_distance_in) == i) {
+            current_distance++;
+            increase_distance_in = reachable_states.size();
+        }
+        for (int successor : graph[reachable_states[i]]) {
+            if (distances[successor] > current_distance) {
+                distances[successor] = current_distance;
+                reachable_states.push_back(successor);
+            }
+        }
     }
 }
-
 
 
 #endif
